+++ lib/es5/sync.d.ts
@@ -5,7 +5,7 @@ export = transform
 
 // transform(records, [options], handler)
 
-type Handler = (record: Array<any>) => any
-declare function transform(records: Array<any>, handler: Handler): Array<any>
-declare function transform(records: Array<any>, options: streamTransform.Options, handler: Handler): Array<any>
-declare namespace transform {}
+type Handler<T = any, U = any> = (record: T) => U
+declare function transform<T = any, U = any>(records: Array<T>, handler: Handler<T, U>): Array<U>
+declare function transform<T = any, U = any>(records: Array<T>, options: streamTransform.Options, handler: Handler<T, U>): Array<U>
+declare namespace transform { }